+++ src/components/manage/AnchorPlugin/components/LinkButton/AddLinkForm.jsx
@@ -14,6 +14,7 @@ import { doesNodeContainClick } from 'semantic-ui-react/dist/commonjs/lib';
 import { defineMessages, injectIntl } from 'react-intl';
 
 import { resetSearchContent, searchContent } from '../../../../../actions';
+import { addAppURL } from '../../../../../helpers';
 import URLUtils from '../../utils/URLUtils';
 
 const messages = defineMessages({
@@ -236,7 +237,7 @@ class AddLinkForm extends Component {
             <li style={{ padding: '5px' }}>
               <button
                 style={{ cursor: 'pointer' }}
-                onClick={e => this.onSelectItem(e, item['@id'])}
+                onClick={e => this.onSelectItem(e, addAppURL(item['@id']))}
                 title={item['@id']}
                 role="link"
               >