+++ src/actions/index.js
@@ -28,9 +28,7 @@ export {
   updateControlpanel,
 } from '@plone/volto/actions/controlpanels/controlpanels';
 export { getDiff } from '@plone/volto/actions/diff/diff';
-export {
-  emailNotification,
-} from '@plone/volto/actions/emailNotification/emailNotification';
+export { emailNotification } from '@plone/volto/actions/emailNotification/emailNotification';
 export {
   createGroup,
   deleteGroup,
@@ -83,3 +81,4 @@ export {
   transitionWorkflow,
 } from '@plone/volto/actions/workflow/workflow';
 export { getQuerystring } from './querystring/querystring';
+export { setSidebarTab } from './sidebar/sidebar';