+++ src/components/manage/Display/Display.jsx
@@ -65,6 +65,10 @@ const customSelectStyles = {
     boxShadow: 'none',
     borderBottom: '2px solid #b8c6c8',
   }),
+  menuList: (styles, state) => ({
+    ...styles,
+    maxHeight: '400px',
+  }),
   indicatorSeparator: styles => ({
     ...styles,
     width: null,