+++ src/server.jsx
@@ -17,6 +17,7 @@ import routes from '~/routes';
 import nlLocale from '~/../locales/nl.json';
 import deLocale from '~/../locales/de.json';
 import enLocale from '~/../locales/en.json';
+import jaLocale from '~/../locales/ja.json';
 
 import {
   Html,
@@ -39,6 +40,7 @@ const locales = {
   en: enLocale,
   nl: nlLocale,
   de: deLocale,
+  ja: jaLocale,
 };
 
 const server = express();