+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 - Remove "documentDescription" class in table block @sverbois
 - General toolbar CSS fixes @sneridagh
 
+### Internal
+
+- Updated to react-select v3 @robdayz
+
 ## 4.0.0-alpha.17 (2019-12-03)
 
 ### Internal