+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added CTRL+ENTER feature in text blocks by default. It creates a newline inside the same text chunk (`<p>`) @sneridagh
 - Automatically switch sidebar on block change @robgietema
 - Fix file and link redirect views @robgietema
+- Japanese translation @terapyon
 
 ### Changes
 
@@ -90,8 +91,6 @@
 ### Added
 
 - Add Node 12 support @timo
-- Japanese translation @terapyon
-
 
 ### Changes