+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 - Added CTRL+ENTER feature in text blocks by default. It creates a newline inside the same text chunk (`<p>`) @sneridagh
 - Automatically switch sidebar on block change @robgietema
+- Fix file and link redirect views @robgietema
 
 ### Changes
 
 - Remove "documentDescription" class in table block @sverbois
-- General toolbar CSS fixes @sneridagh
+- General toolbar more and personal tools menu CSS fixes @sneridagh
+- Fix bug that lead to crashing the view when deleting the last row of a table
 
 ### Internal