+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added CTRL+ENTER feature in text blocks by default. It creates a newline inside the same text chunk (`<p>`) @sneridagh
 - Automatically switch sidebar on block change @robgietema
+- Add new line in text blocks tests @rodfersou
 
 ### Changes