+++ CHANGELOG.md
@@ -7,11 +7,14 @@
 - Added CTRL+ENTER feature in text blocks by default. It creates a newline inside the same text chunk (`<p>`) @sneridagh
 - Automatically switch sidebar on block change @robgietema
 - Add new line in text blocks tests @rodfersou
+- Fix file and link redirect views @robgietema
+- Japanese translation @terapyon
 
 ### Changes
 
 - Remove "documentDescription" class in table block @sverbois
-- General toolbar CSS fixes @sneridagh
+- General toolbar more and personal tools menu CSS fixes @sneridagh
+- Fix bug that lead to crashing the view when deleting the last row of a table
 
 ## 4.0.0-alpha.17 (2019-12-03)