+++ CHANGELOG.md
@@ -6,9 +6,9 @@
 
 - Added CTRL+ENTER feature in text blocks by default. It creates a newline inside the same text chunk (`<p>`) @sneridagh
 - Automatically switch sidebar on block change @robgietema
-- Add new line in text blocks tests @rodfersou
 - Fix file and link redirect views @robgietema
 - Japanese translation @terapyon
+- Add new line in text blocks tests @rodfersou
 
 ### Changes