+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Changes
 
 - Remove "documentDescription" class in table block @sverbois
+- Added possibility to work with vimeo-videos instead of youtube-videos in the video block @wkbkhard
+- Fixed Issue 1021: typing in a "wrong" URL leads to error @wkbkhard
 - General toolbar more and personal tools menu CSS fixes @sneridagh
 - Fix bug that lead to crashing the view when deleting the last row of a table
 - Fix Select widget bug if the field has already the options in the `choices` schema, do not trigger the vocabulary request @sneridagh